requirements.txt
altair==5.4.1
annotated-types==0.7.0
asn1crypto==1.5.1
atpublic==5.0
attrs==24.2.0
blinker==1.8.2
boto3==1.35.77
botocore==1.35.77
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
cryptography==43.0.1
et_xmlfile==2.0.0
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.9.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
protobuf==5.28.2
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
pymsteams==0.2.5
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
# pywin32==306
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
s3transfer==0.10.4
six==1.17.0
smmap==5.0.1
snowflake-connector-python==3.12.2
snowflake-snowpark-python==1.22.1
sortedcontainers==2.4.0
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
watchdog==4.0.2
setuptools>=75.1.0
